Use double instead of float
Java uses the double as the default data type for floating point numbers.
The suffix f can therefore be omitted.

+++ src/main/java/rocks/friedrich/engine_omega/FrameUpdateListener.java
@@ -16,5 +16,5 @@ public interface FrameUpdateListener
      *                     Update vergangen
      */
     @API
-    void onFrameUpdate(float deltaSeconds);
+    void onFrameUpdate(double deltaSeconds);
 }
