+++ src/main/java/rocks/friedrich/engine_omega/actor/Image.java
@@ -41,9 +41,9 @@ public class Image extends Actor
      */
     private final BufferedImage image;
 
-    private float width;
+    private double width;
 
-    private float height;
+    private double height;
 
     private boolean flipVertical = false;
 
@@ -57,7 +57,7 @@ public class Image extends Actor
      * @param height   Die Höhe in M für das Bildobjekt
      */
     @API
-    public Image(String filepath, float width, float height)
+    public Image(String filepath, double width, double height)
     {
         super(() -> FixtureBuilder.createSimpleRectangularFixture(width,
                 height));
@@ -77,7 +77,7 @@ public class Image extends Actor
      *                      der Engine entsprechen.
      */
     @API
-    public Image(String filepath, final float pixelPerMeter)
+    public Image(String filepath, final double pixelPerMeter)
     {
         super(() -> FixtureBuilder.createSimpleRectangularFixture(
                 ImageLoader.load(filepath).getWidth() / pixelPerMeter,
@@ -98,7 +98,7 @@ public class Image extends Actor
      *                      der Engine entsprechen.
      */
     @API
-    public Image(BufferedImage image, final float pixelPerMeter)
+    public Image(BufferedImage image, final double pixelPerMeter)
     {
         super(() -> FixtureBuilder.createSimpleRectangularFixture(
                 image.getWidth() / pixelPerMeter,
@@ -135,9 +135,9 @@ public class Image extends Actor
      * @param width  Die neue Breite des Objekts in M.
      * @param height Die neue Höhe des Objekts in M.
      *
-     * @see #resetPixelPerMeter(float)
+     * @see #resetPixelPerMeter(double)
      */
-    public void resetImageSize(float width, float height)
+    public void resetImageSize(double width, double height)
     {
         assertViableSizes(width, height);
         this.width = width;
@@ -155,16 +155,16 @@ public class Image extends Actor
      *                      an, wie viele Pixel in der Bilddatei einem Meter in
      *                      der Engine entsprechen.
      *
-     * @see #resetImageSize(float, float)
+     * @see #resetImageSize(double, double)
      */
-    public void resetPixelPerMeter(float pixelPerMeter)
+    public void resetPixelPerMeter(double pixelPerMeter)
     {
         assertViablePPM(pixelPerMeter);
         resetImageSize(image.getWidth() / pixelPerMeter,
                 image.getHeight() / pixelPerMeter);
     }
 
-    private void assertViableSizes(float width, float height)
+    private void assertViableSizes(double width, double height)
     {
         if (width <= 0 || height <= 0)
         {
@@ -173,7 +173,7 @@ public class Image extends Actor
         }
     }
 
-    private void assertViablePPM(float pixelPerMeter)
+    private void assertViablePPM(double pixelPerMeter)
     {
         if (pixelPerMeter <= 0)
         {
@@ -241,7 +241,7 @@ public class Image extends Actor
      * {@inheritDoc}
      */
     @Override
-    public void render(Graphics2D g, float pixelPerMeter)
+    public void render(Graphics2D g, double pixelPerMeter)
     {
         AffineTransform pre = g.getTransform();
         g.scale(width * pixelPerMeter / this.image.getWidth(),
