+++ src/main/java/rocks/friedrich/engine_omega/actor/StatefulAnimation.java
@@ -42,13 +42,13 @@ public class StatefulAnimation<State> extends Actor
 
     private AnimationFrame[] currentAnimation = null;
 
-    private float currentTime = 0;
+    private double currentTime = 0;
 
     private int currentIndex = 0;
 
-    private float width;
+    private double width;
 
-    private float height;
+    private double height;
 
     private boolean flipHorizontal = false;
 
@@ -56,7 +56,7 @@ public class StatefulAnimation<State> extends Actor
 
     private boolean animationPaused = false;
 
-    public StatefulAnimation(float width, float height)
+    public StatefulAnimation(double width, double height)
     {
         super(() -> FixtureBuilder.createSimpleRectangularFixture(width,
                 height));
@@ -66,13 +66,13 @@ public class StatefulAnimation<State> extends Actor
     }
 
     @API
-    public float getWidth()
+    public double getWidth()
     {
         return width;
     }
 
     @API
-    public float getHeight()
+    public double getHeight()
     {
         return height;
     }
@@ -311,7 +311,7 @@ public class StatefulAnimation<State> extends Actor
      *                      bis der Frame gewechselt wird.
      */
     @API
-    public void setFrameDuration(State state, float frameDuration)
+    public void setFrameDuration(State state, double frameDuration)
     {
         if (!states.containsKey(state))
         {
@@ -328,7 +328,7 @@ public class StatefulAnimation<State> extends Actor
      * Methode wird frameweise über einen anonymen Listener aufgerufen.
      */
     @Internal
-    private void internalOnFrameUpdate(float frameDuration)
+    private void internalOnFrameUpdate(double frameDuration)
     {
         if (currentAnimation == null || currentAnimation.length == 0
                 || animationPaused)
@@ -365,7 +365,7 @@ public class StatefulAnimation<State> extends Actor
      * @param height Neue Höhe für das Rechteck.
      */
     @API
-    public void setSize(float width, float height)
+    public void setSize(double width, double height)
     {
         assertPositiveWidthAndHeight(width, height);
         this.width = width;
@@ -376,7 +376,7 @@ public class StatefulAnimation<State> extends Actor
 
     @Internal
     @Override
-    public void render(Graphics2D g, float pixelPerMeter)
+    public void render(Graphics2D g, double pixelPerMeter)
     {
         if (currentAnimation == null || currentAnimation.length == 0)
         {
